+++ src/components/SubwaySelect/SubwayListModal.tsx
@@ -34,7 +34,7 @@ const SubwayListModal = ({ onClose, onSelect }: SubwayListModalProps) => {
 
   return (
     <LayoutRootPortal>
-      <div className='absolute left-0 top-0 z-10 flex h-full w-full flex-col bg-white'>
+      <div className='absolute left-0 top-0 z-10 flex h-full w-full flex-col bg-gray-bg-base'>
         <div className='flex items-center gap-2 border-b border-gray-accent7 px-4 py-2'>
           <Button onClick={onClose} className='h-fit w-fit'>
             <Icon
