+++ src/constants/times.ts
@@ -1,28 +1,1 @@
 export const TIMES = ['평일 오전', '평일 오후', '주말 오전', '주말 오후'];
-export const TIME_PICKER_OPTIONS = [
-  '오전 10:00',
-  '오전 10:30',
-  '오전 11:00',
-  '오전 11:30',
-  '오후 12:00',
-  '오후 12:30',
-  '오후 1:00',
-  '오후 1:30',
-  '오후 2:00',
-  '오후 2:30',
-  '오후 3:00',
-  '오후 3:30',
-  '오후 4:00',
-  '오후 4:30',
-  '오후 5:00',
-  '오후 5:30',
-  '오후 6:00',
-  '오후 6:30',
-  '오후 7:00',
-  '오후 7:30',
-  '오후 8:00',
-  '오후 8:30',
-  '오후 9:00',
-  '오후 9:30',
-  '오후 10:00',
-];
